renamed to azurefunctions-extensions-base
