+++ azurefunctions-extensions-base/MANIFEST.in
