+++ azurefunctions-extensions-base/azurefunctions/extensions/base/meta.py
@@ -4,7 +4,7 @@
 import abc
 import inspect
 import json
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Dict, Mapping, Optional, Tuple, Union
 
 from . import sdkType, utils
