+++ azurefunctions-extensions-base/azurefunctions/extensions/base/utils.py
