+++ azurefunctions-extensions-base/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools >= 61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "azurefunctions-extension-base"
+name = "azurefunctions-extensions-base"
 dynamic = ["version"]
 requires-python = ">=3.8"
 authors = [{ name = "Azure Functions team at Microsoft Corp.", email = "azurefunctions@microsoft.com"}]
@@ -35,8 +35,8 @@ dev = [
     ]
 
 [tool.setuptools.dynamic]
-version = {attr = "azurefunctions.extension.base.__version__"}
+version = {attr = "azurefunctions.extensions.base.__version__"}
 
 [tool.setuptools.packages.find]
-exclude = ['azurefunctions.extension', 'azurefunctions', 'tests']
+exclude = ['azurefunctions.extensions', 'azurefunctions', 'tests']
