+++ azurefunctions-extensions-base/tests/test_utils.py
@@ -4,7 +4,7 @@
 import unittest
 from abc import ABC
 
-from azurefunctions.extension.base import meta, sdkType, utils
+from azurefunctions.extensions.base import meta, sdkType, utils
 
 
 class MockParamTypeInfo:
