+++ azurefunctions-extensions-base/tests/test_web.py
@@ -1,7 +1,6 @@
 import unittest
-from unittest.mock import MagicMock, patch
 
-from azurefunctions.extension.base import (
+from azurefunctions.extensions.base import (
     HttpV2FeatureChecker,
     ModuleTrackerMeta,
     RequestSynchronizer,
